% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand-type.R
\name{adi}
\alias{adi}
\title{https://deep-and-shallow.com/2020/10/07/forecast-error-measures-intermittent-demand/
Helper functions
Consider moving them into a separate package}
\usage{
adi(x, ...)
}
\description{
https://deep-and-shallow.com/2020/10/07/forecast-error-measures-intermittent-demand/
Helper functions
Consider moving them into a separate package
}
