% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand-type.R
\name{leadtrail}
\alias{leadtrail}
\title{Taken from https://github.com/trnnick/tsutils/blob/master/R/leadtrail.R}
\usage{
leadtrail(
  x,
  rm = c("zeros", "na"),
  lead = c(TRUE, FALSE),
  trail = c(TRUE, FALSE)
)
}
\arguments{
\item{x}{vector of values to check.}

\item{rm}{what to remove, can be \code{"zeros"} or \code{"na"}.}

\item{lead}{If \code{TRUE}, then leading values are removed.}

\item{trail}{If \code{TRUE}, then trailing values are removed.}
}
\value{
Resulting vector.
}
\description{
Remove leading/training zeros/NAs
}
\details{
Remove leading or trailing zeros or NAs from a vector.
}
\examples{
x <- c(rep(0,5),rnorm(100),rep(0,5))
leadtrail(x)

}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}.
}
