% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5-package.R
\docType{package}
\name{m5-package}
\alias{m5-package}
\alias{m5}
\title{m5: M5 Walmart Challenge Data}
\description{
Constains functions, which facilitate downloading, loading and preparing data
from M5 Walmart Challenge (by University of Nicosia, hosted on Kaggle).
It also includes a ready-to-use built-in M5 subset named \code{tiny_m5}.
}
\author{
Krzysztof Joachimiak
}
\keyword{internal}
\keyword{package}
