% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{m5_get_raw_*}
\alias{m5_get_raw_*}
\title{Load raw CSV files using data.table::fread function}
\arguments{
\item{path}{The directory with the unzipped M5 data files}
}
\value{
The function returns a list of five data.tables:

* sales_train (evaluation/validation)
* sales_test (evaluation/validation)
* sell_prices
* calendar
* weights (evaluation/validation)
}
\description{
Load raw CSV files using data.table::fread function
}
\examples{
\dontrun{
library(m5)
library(zeallot)

m5_download('data')
c(sales_train,
  sales_test,
  sell_prices,
  calendar,
  ) \%<-\% m5_get_raw_evaluation('data')
}
}
\references{
[m5-forecasts repo by Nixtla](https://github.com/Nixtla/m5-forecasts)
}
