% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{m5_prepare}
\alias{m5_prepare}
\title{Prepare the ready-to-use M5 data in one data.frame}
\usage{
m5_prepare(sales_train, sales_test, calendar, sell_prices)
}
\arguments{
\item{sales_train}{A data.frame with M5 train data}

\item{sales_test}{A data.frame with M5 test data}

\item{calendar}{A data.frame with M5 calendar}

\item{sell_prices}{A data.frame with M5 sell_prices}
}
\description{
It's a memory-efficient function, which uses \code{data.table} under the hood.
However, it still not recommended to use this function on PCs with < 16GB RAM.
In such case, consider to use a custom solution
based on \verb{[arrow](https://arrow.apache.org/docs/r/)} or \verb{[disk.frame](https://diskframe.com/index.html)}
}
\examples{
library(m5)
library(zeallot)
\dontrun{
m5_download('data')
c(sales_train,
  sales_test,
  sell_prices,
  calendar) \%<-\% m5_get_raw_evaluation('data')

m5_data  <-
   m5_prepare(sales_train, sales_test, calendar, sell_prices)
}

}
