% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m5-package.R
\docType{package}
\name{m5-package}
\alias{m5-package}
\alias{m5}
\title{m5: 'M5 Forecasting' Challenges Data}
\description{
Contains functions, which facilitate downloading, loading and preparing data
from 'M5 Forecasting' challenges (by 'University of Nicosia', hosted on 'Kaggle').
The data itself is set of time series of different product sales in 'Walmart'.
The package also includes a ready-to-use built-in M5 subset named \code{tiny_m5}.
For detailed information about the challenges, see:
Makridakis, Spyros & Spiliotis, Evangelos & Assimakopoulos, Vassilis. (2020).
The M5 Accuracy competition: Results, findings and conclusions. \url{doi:10.1016/j.ijforecast.2021.10.009}
}
\author{
Krzysztof Joachimiak
}
\keyword{internal}
\keyword{package}
