% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand-type.R
\name{m5_demand_type}
\alias{m5_demand_type}
\title{Classify time series of the particular items}
\usage{
m5_demand_type(data)
}
\arguments{
\item{data}{The result of the `m5_prepare` function.}
}
\value{
A `data.frame` containing item ids, ADI and CV2 scores as well as the final
class chosen based on the aforementioned scores.
}
\description{
Each time series in the dataset can be assigned one of the following classes:
}
\details{
* Intermittent
* Lumpy
* Smooth
* Erratic
}
\references{

}
