% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{m5_download}
\alias{m5_download}
\title{Download and unzip the raw data to the specified directory}
\usage{
m5_download(path, unzip = TRUE)
}
\arguments{
\item{path}{A directory name to save the zip file}

\item{unzip}{Automatically uznip the file when the downloading is finished.
Default: TRUE. The \code{exdir} argument in the \code{unzip} function is the directory
name the file was downloaded into.}
}
\value{
Returns nothing; the result of the function call is the \code{m5.zip} file
downloading and extraction.
}
\description{
Download and unzip the raw data to the specified directory
}
\note{
If you struggle with timeout problems, please increase the timeout value
using \verb{options(timeout=<new_timeout_value>)}
}
\examples{
\donttest{
m5_download('data')
}

}
\references{
\itemize{
\item \href{https://www.kaggle.com/c/m5-forecasting-accuracy}{M5 Forecasting - Accuracy}
\item \href{https://www.kaggle.com/c/m5-forecasting-uncertainty}{M5 Forecasting - Uncertainty}
}
}
