% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{m5_download}
\alias{m5_download}
\title{Download and unzip the raw data to the specified directory}
\usage{
m5_download(path, unzip = TRUE)
}
\arguments{
\item{path}{A directory name to save the zip file}

\item{unzip}{Automatically uznip the file when the downloading is finished.
Default: TRUE. The `exdir` argument in the `unzip` function is the directory
name the file was downloaded into.}
}
\description{
Download and unzip the raw data to the specified directory
}
\examples{
\dontrun{
m5_download('data')
}

}
\references{
* [M5 Forecasting - Accuracy](https://www.kaggle.com/c/m5-forecasting-accuracy)
* [M5 Forecasting - Uncertainty](https://www.kaggle.com/c/m5-forecasting-uncertainty)
}
